# backend/openapi.snapshot.json
{
  "components": {
    "schemas": {
      "ApprovalResumeRequest": {
        "properties": {
          "approval_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Approval Id",
            "type": "string"
          },
          "decision": {
            "default": "approve",
            "enum": [
              "approve",
              "reject"
            ],
            "title": "Decision",
            "type": "string"
          },
          "execution_ref": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExecutionRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "thread_id": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Id"
          }
        },
        "required": [
          "approval_id"
        ],
        "title": "ApprovalResumeRequest",
        "type": "object"
      },
...
      "DefaultProviderUpdate": {
        "properties": {
          "model": {
            "title": "Model",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "model"
        ],
        "title": "DefaultProviderUpdate",
        "type": "object"
      },
...
      "ExecutionRef": {
        "description": "Server-issued identity of one interrupted graph execution.",
        "properties": {
          "agent_id": {
            "minLength": 1,
            "title": "Agent Id",
            "type": "string"
          },
          "checkpoint_ns": {
            "title": "Checkpoint Ns",
            "type": "string"
          },
          "interrupt_id": {
            "minLength": 1,
            "title": "Interrupt Id",
            "type": "string"
          },
          "interrupt_type": {
            "minLength": 1,
            "title": "Interrupt Type",
            "type": "string"
          },
          "thread_id": {
            "minLength": 1,
            "title": "Thread Id",
            "type": "string"
          }
        },
        "required": [
          "agent_id",
          "thread_id",
          "checkpoint_ns",
          "interrupt_id",
          "interrupt_type"
        ],
        "title": "ExecutionRef",
        "type": "object"
      },
...
      "GuardConfigUpdate": {
        "properties": {
          "level": {
            "title": "Level",
            "type": "string"
          }
        },
        "required": [
          "level"
        ],
        "title": "GuardConfigUpdate",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProviderConfigUpdate": {
        "properties": {
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          }
        },
        "title": "ProviderConfigUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "钻井行业 DeepAgents 智能体系统 - 自主规划与执行",
    "title": "BoetClaw DeepAgents",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/providers/default": {
      "put": {
        "operationId": "update_default_provider_api_v1_providers_default_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DefaultProviderUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Default Provider",
        "tags": [
          "Providers"
        ]
      }
    },
    "/api/v1/providers/{name}/check": {
      "post": {
        "operationId": "check_connection_api_v1_providers__name__check_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Check Connection",
        "tags": [
          "Providers"
        ]
      }
    },
    "/api/v1/providers/{name}/config": {
      "get": {
        "operationId": "get_provider_config_api_v1_providers__name__config_get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Provider Config",
        "tags": [
          "Providers"
        ]
      },
      "put": {
        "operationId": "update_provider_config_api_v1_providers__name__config_put",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderConfigUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Provider Config",
        "tags": [
          "Providers"
        ]
      }
    },
    "/api/v1/providers/{name}/models": {
      "get": {
        "operationId": "list_models_api_v1_providers__name__models_get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Models",
        "tags": [
          "Providers"
        ]
      }
    },
...
    "/api/v1/security/approvals/resume": {
      "post": {
        "operationId": "resume_approval_api_v1_security_approvals_resume_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalResumeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resume Approval",
        "tags": [
          "Security"
        ]
      }
    },
    "/api/v1/security/config": {
      "get": {
        "operationId": "get_config_api_v1_security_config_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Config",
        "tags": [
          "Security"
        ]
      },
      "put": {
        "operationId": "update_config_api_v1_security_config_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuardConfigUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Config",
        "tags": [
          "Security"
        ]
      }
    },
...
  }
}